1. 修改需求

+++ src/main/java/com/hanzoy/yuekewei/pojo/dto/result/GetCourseInfoResult.java
@@ -22,7 +22,7 @@ public class GetCourseInfoResult {
     /**
      * 剩余课时数量
      */
-    private Integer surplusCount;
+    private Double surplusCount;
 
     /**
      * timetable Id
